requirements.txt
# API server
fastapi>=0.111
uvicorn[standard]>=0.29
python-multipart>=0.0.9
pydantic>=2.7

# ML core
torch>=2.3
torchvision>=0.18
transformers>=4.41
peft>=0.11
accelerate>=0.30
numpy>=1.26
Pillow>=10.3
scipy>=1.13
scikit-learn>=1.4

# VQASynth pipeline (depth, localization, scene fusion, VQA generation)
# Install with: pip install git+https://github.com/remyxai/VQASynth.git
vqasynth @ git+https://github.com/remyxai/VQASynth.git

# VQASynth dependencies
open3d>=0.18
onnxruntime==1.15.1
depth-pro @ git+https://github.com/apple/ml-depth-pro.git
spacy>=3.7
sam2>=0.3

# Pipeline / annotation
opencv-python-headless>=4.9

# LLM for QA generation / CoT reasoning
anthropic>=0.28

# Utilities
tqdm>=4.66
